refactor: start refactoring ad with new snapshots
