+++ packages/ad/__tests__/android/ad-placeholder-config.android.test.js
@@ -1,0 +1,5 @@
+import shared from "../ad-placeholder-config.shared";
+
+describe("android", () => {
+  shared();
+});
