+++ packages/ad/__tests__/android/ad-placeholder.android.test.js
@@ -1,7 +1,5 @@
-import sharedNative from "../ad-placeholder.native";
-import shared from "../ad-placeholder.shared";
+import shared from "../ad-placeholder.native";
 
 describe("android", () => {
-  sharedNative();
   shared();
 });
