+++ packages/ad/__tests__/ios/ad-placeholder-config.ios.test.js
@@ -1,0 +1,5 @@
+import shared from "../ad-placeholder-config.shared";
+
+describe("ios", () => {
+  shared();
+});
