+++ packages/ad/__tests__/web/ad-placeholder-config.web.test.js
@@ -1,0 +1,5 @@
+import shared from "../ad-placeholder-config.shared";
+
+describe("web", () => {
+  shared();
+});
